+++ energy-consumption/src/test/java/com/powerhouse/interview/util/ValidateTest.java
@@ -41,19 +41,19 @@ public class ValidateTest {
 	@Test
 	public void consumptionsShouldBeConsistenWithFractions() {
 		
-		Map<Month, Integer> meterReadingMap = new TreeMap<Month, Integer>(new MonthComparator());
-		meterReadingMap.put(Month.JANUARY, 0);
-		meterReadingMap.put(Month.FEBRUARY, 0);
-		meterReadingMap.put(Month.MARCH, 0);
-		meterReadingMap.put(Month.APRIL, 0);
-		meterReadingMap.put(Month.MAY, 0);
-		meterReadingMap.put(Month.JUNE, 0);
-		meterReadingMap.put(Month.JULY, 0);
-		meterReadingMap.put(Month.AUGUST, 0);
-		meterReadingMap.put(Month.SEPTEMBER, 0);
-		meterReadingMap.put(Month.OCTOBER, 0);
-		meterReadingMap.put(Month.NOVEMBER, 0);
-		meterReadingMap.put(Month.DECEMBER, 10);
+		Map<Month, Double> meterReadingMap = new TreeMap<Month, Double>(new MonthComparator());
+		meterReadingMap.put(Month.JANUARY, 0d);
+		meterReadingMap.put(Month.FEBRUARY, 0d);
+		meterReadingMap.put(Month.MARCH, 0d);
+		meterReadingMap.put(Month.APRIL, 0d);
+		meterReadingMap.put(Month.MAY, 0d);
+		meterReadingMap.put(Month.JUNE, 0d);
+		meterReadingMap.put(Month.JULY, 0d);
+		meterReadingMap.put(Month.AUGUST, 0d);
+		meterReadingMap.put(Month.SEPTEMBER, 0d);
+		meterReadingMap.put(Month.OCTOBER, 0d);
+		meterReadingMap.put(Month.NOVEMBER, 0d);
+		meterReadingMap.put(Month.DECEMBER, 10d);
 		
 		Map<Month, Double> fractionMap = new TreeMap<Month, Double>(new MonthComparator());
 		fractionMap.put(Month.JANUARY, 0.0);
@@ -74,50 +74,50 @@ public class ValidateTest {
 	
 	@Test
 	public void meterReadingsCanNotBeLowerThanPreviousMonth() {
-		Map<Month, Integer> map = new HashMap<Month, Integer>();
+		Map<Month, Double> map = new HashMap<Month, Double>();
 		
-		map.put(Month.JANUARY, 10);
-		map.put(Month.FEBRUARY, 9);
-		map.put(Month.MARCH, 9);
-		map.put(Month.APRIL, 9);
-		map.put(Month.MAY, 9);
-		map.put(Month.JUNE, 9);
-		map.put(Month.JULY, 9);
-		map.put(Month.AUGUST, 9);
-		map.put(Month.SEPTEMBER, 9);
-		map.put(Month.OCTOBER, 9);
-		map.put(Month.NOVEMBER, 9);
-		map.put(Month.DECEMBER, 9);
+		map.put(Month.JANUARY, 10d);
+		map.put(Month.FEBRUARY, 9d);
+		map.put(Month.MARCH, 9d);
+		map.put(Month.APRIL, 9d);
+		map.put(Month.MAY, 9d);
+		map.put(Month.JUNE, 9d);
+		map.put(Month.JULY, 9d);
+		map.put(Month.AUGUST, 9d);
+		map.put(Month.SEPTEMBER, 9d);
+		map.put(Month.OCTOBER, 9d);
+		map.put(Month.NOVEMBER, 9d);
+		map.put(Month.DECEMBER, 9d);
 		assertFalse(validate.meterReading(map));
 		
-		map.put(Month.FEBRUARY, 10);
+		map.put(Month.FEBRUARY, 10d);
 		assertFalse(validate.meterReading(map));
 		
-		map.put(Month.MARCH, 10);
+		map.put(Month.MARCH, 10d);
 		assertFalse(validate.meterReading(map));
 		
-		map.put(Month.APRIL, 10);
+		map.put(Month.APRIL, 10d);
 		assertFalse(validate.meterReading(map));
 		
-		map.put(Month.MAY, 10);
+		map.put(Month.MAY, 10d);
 		assertFalse(validate.meterReading(map));
 		
-		map.put(Month.JUNE, 10);
+		map.put(Month.JUNE, 10d);
 		assertFalse(validate.meterReading(map));
 		
-		map.put(Month.JULY, 10);
+		map.put(Month.JULY, 10d);
 		assertFalse(validate.meterReading(map));
 		
-		map.put(Month.AUGUST, 10);
+		map.put(Month.AUGUST, 10d);
 		assertFalse(validate.meterReading(map));
 		
-		map.put(Month.SEPTEMBER, 10);
+		map.put(Month.SEPTEMBER, 10d);
 		assertFalse(validate.meterReading(map));
 		
-		map.put(Month.OCTOBER, 10);
+		map.put(Month.OCTOBER, 10d);
 		assertFalse(validate.meterReading(map));
 		
-		map.put(Month.NOVEMBER, 10);
+		map.put(Month.NOVEMBER, 10d);
 		assertFalse(validate.meterReading(map));
 		
 	}
